refactor of code, now modules in core/api/ are consistent with docs
