+++ core/api/cart/submodules/closeMicrocart.js
@@ -2,14 +2,14 @@
  * [Helper] Optional UI functionality for closing microcart.
  *
  * #### Methods
- * - **`closeMicrocart`** sets `microcart` and `overlay` property from ui state to `false`. Dispatches `ui/setMicrocart'` Vuex action
+ * - **`closeMicrocart`** sets `microcart` and `overlay` property from submodules state to `false`. Dispatches `submodules/setMicrocart'` Vuex action
  *
  * Part of [Cart API Module](https://github.com/DivanteLtd/vue-storefront/tree/master/doc/api-modules)
  */
 export const closeMicrocart = {
   methods: {
     closeMicrocart () {
-      this.$store.commit('ui/setMicrocart', false)
+      this.$store.commit('submodules/setMicrocart', false)
     }
   }
 }
