+++ core/api/cart/submodules/isMicrocartOpen.js
@@ -2,7 +2,7 @@
  * [Helper] Optional UI functionality for informing if microcart is open.
  *
  * #### Computed properties
- * - **`isMicrocartOpen`** returns `true` if Microcart is open. Retuens `ui/microcart` Vuex state.
+ * - **`isMicrocartOpen`** returns `true` if Microcart is open. Retuens `submodules/microcart` Vuex state.
  *
  * Part of [Cart API Module](https://github.com/DivanteLtd/vue-storefront/tree/master/doc/api-modules)
  */
